some updates
Alyssa: did you use any outside jar files? Because  PDFtoImage is
causing a problem for me.
Shawn: I got the DatabaseImporter and MultipleChoiceQuestion working
together. There is one issue now is ExamGenerator with creating an Exam
and in HTMLGenerator GenerateQRCode isn't passing the right things.
I added a comments in the code I think for most of them I added <co> to
the end of the comments to help distinguish them