GUI
a VERY simple gui which took way too long to get sort of working.
We should really resolve the "how to save correct choices" problem.
I have it as being answer_id which only works for the first 4 questions.
BUT should work after that issue is resolved.
I added a test_id column to the responses table because it was the only
way to just get the choices for one test.