Javascript QR codes
Unfortunately, they don't scale to their minimum size and become
extremely pixelated. I've included the JSOutput folder which has
samples. I also added an additional method to QRCodeHandler so the
contents of the QR codes are generated by a method that can be accessed
elsewhere. This method is now also called in the generation method where
the contents are determined. Sorry bout it.